\name{gWidgetsRGtk2-undocumented}
\alias{gWidgetsRGtk2-undocumented}
%% these are by hand
\alias{editSubsetDialog}
\alias{editSelectDialog}
%% These are found using FindAllGenerics.R
\alias{[-methods}
\alias{[,gWidgetRGtk-method}
\alias{[,gCheckboxRGtk-method}
\alias{[,gCheckboxgroupRGtk-method}
\alias{[,gCheckboxgroupTableRGtk-method}
\alias{[,gCommandlineRGtk-method}
\alias{[,gDfEditRGtk-method}
\alias{[,gDroplistRGtk-method}
\alias{[,gEditRGtk-method}
\alias{[,GtkTreeView-method}
\alias{[,gGridRGtk-method}
\alias{[,gLayoutRGtk-method}
\alias{[,gMenuRGtk-method}
\alias{[,gNotebookRGtk-method}
\alias{[,gRadioRGtk-method}
\alias{[,RGtkDataFrame-method}
\alias{[,gToolbarRGtk-method}
\alias{[,gTreeRGtk-method}
\alias{[,gVarbrowserRGtk-method}
\alias{.delete,gNotebookRGtk,guiWidgetsToolkitRGtk2,ANY-method}
\alias{.delete,gToolbarRGtk,guiWidgetsToolkitRGtk2,ANY-method}

\alias{.svalue<-,gNotebookRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gPanedgroupRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gRadioRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gSliderRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gSpinbuttonRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gStatusbarRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gTextRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gToolbarRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gTreeRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gAddargRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gBivariateRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gModelRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gWindowRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gExpandgroupRGtk,guiWidgetsToolkitRGtk2,ANY,numeric-method}
%%
%%
%% --  .glabel --
%%
\alias{.glabel-methods}
\alias{.glabel,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gbutton --
%%
\alias{.gbutton-methods}
\alias{.gbutton,guiWidgetsToolkitRGtk2-method} 
\alias{.gbutton,guiWidgetsToolkitRGtk2,ANY,ANY,ANY,ANY-method}
\alias{.gbutton,guiWidgetsToolkitRGtk2,ANY,ANY,ANY,gActionRGtk-method}	 
\alias{.gbutton,guiWidgetsToolkitRGtk2,ANY,ANY,ANY,guiComponent-method} 

%%
%% --  .gcheckbox --
%%
\alias{.gcheckbox-methods}
\alias{.gcheckbox,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gradio --
%%
\alias{.gradio-methods}
\alias{.gradio,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gdroplist --
%%
\alias{.gdroplist-methods}
\alias{.gdroplist,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gcheckboxgroup --
%%
\alias{.gcheckboxgroup-methods}
\alias{.gcheckboxgroup,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gspinbutton --
%%
\alias{.gspinbutton-methods}
\alias{.gspinbutton,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gslider --
%%
\alias{.gslider-methods}
\alias{.gslider,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gedit --
%%
\alias{.gedit-methods}
\alias{.gedit,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gtext --
%%
\alias{.gtext-methods}
\alias{.gtext,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gaction --
%%
\alias{.gaction-methods}
\alias{.gaction,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gmenu --
%%
\alias{.gmenu-methods}
\alias{.gmenu,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gtoolbar --
%%
\alias{.gtoolbar-methods}
\alias{.gtoolbar,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gtable --
%%
\alias{.gtable-methods}
\alias{.gtable,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gdf --
%%
\alias{.gdf-methods}
\alias{.gdf,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gdfedit --
%%
\alias{.gdfedit-methods}
\alias{.gdfedit,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gdfnotebook --
%%
\alias{.gdfnotebook-methods}
\alias{.gdfnotebook,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gtree --
%%
\alias{.gtree-methods}
\alias{.gtree,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gfile --
%%
\alias{.gfile-methods}
\alias{.gfile,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gfilebrowse --
%%
\alias{.gfilebrowse-methods}
\alias{.gfilebrowse,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gcalendar --
%%
\alias{.gcalendar-methods}
\alias{.gcalendar,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .ggraphics --
%%
\alias{.ggraphics-methods}
\alias{.ggraphics,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .ggraphicsnotebook --
%%
\alias{.ggraphicsnotebook-methods}
\alias{.ggraphicsnotebook,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gimage --
%%
\alias{.gimage-methods}
\alias{.gimage,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gstatusbar --
%%
\alias{.gstatusbar-methods}
\alias{.gstatusbar,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gseparator --
%%
\alias{.gseparator-methods}
\alias{.gseparator,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gcommandline --
%%
\alias{.gcommandline-methods}
\alias{.gcommandline,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .ghelp --
%%
\alias{.ghelp-methods}
\alias{.ghelp,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .ghelpbrowser --
%%
\alias{.ghelpbrowser-methods}
\alias{.ghelpbrowser,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .ggenericwidget --
%%
\alias{.ggenericwidget-methods}
\alias{.ggenericwidget,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gvarbrowser --
%%
\alias{.gvarbrowser-methods}
\alias{.gvarbrowser,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gwindow --
%%
\alias{.gwindow-methods}
\alias{.gwindow,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .ggroup --
%%
\alias{.ggroup-methods}
\alias{.ggroup,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gframe --
%%
\alias{.gframe-methods}
\alias{.gframe,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gexpandgroup --
%%
\alias{.gexpandgroup-methods}
\alias{.gexpandgroup,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gnotebook --
%%
\alias{.gnotebook-methods}
\alias{.gnotebook,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .glayout --
%%
\alias{.glayout-methods}
\alias{.glayout,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gpanedgroup --
%%
\alias{.gpanedgroup-methods}
\alias{.gpanedgroup,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .addStockIcons --
%%
\alias{.addStockIcons-methods}
\alias{.addStockIcons,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .getStockIcons --
%%
\alias{.getStockIcons-methods}
\alias{.getStockIcons,guiWidgetsToolkitRGtk2-method} 
%%
%% --  svalue --
%%
\alias{svalue-methods}
\alias{svalue,guiWidget-method}
\alias{svalue,gWidgetRGtk-method}
\alias{svalue,character-method}
\alias{svalue,GtkEntry-method}
\alias{svalue,gSubsetbyRGtk-method}
\alias{svalue,GtkTreeViewColumn-method} 
%%
%% --  .svalue --
%%
\alias{.svalue-methods}
\alias{.svalue,character,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gActionRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gActionRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gButtonRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCalendarRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCalendarRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCheckboxRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCheckboxgroupRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCheckboxgroupTableRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCommandlineRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gDfEditRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gDfEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gDroplistRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gEditRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,GtkEntry,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gExpandgroupRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,GtkTreeView,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gGridRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gGenericWidgetRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gHelpRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gImageRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gLabelRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gMenuRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gPanedgroupRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gNotebookRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gRadioRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gSliderRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gSpinbuttonRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gStatusbarRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gTextRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gToolbarRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gTreeRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gVarbrowserRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gAddargRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gUnivariateRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gUnivariateTableRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gFileURLRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gBivariateRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gModelRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gLmerRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gWindowRGtk,ANY,ANY,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,character,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gButtonRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCheckboxRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gCommandlineRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gDroplistRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gExpandgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gHelpRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gImageRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gLabelRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gMenuRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gPanedgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gRadioRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gSliderRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gSpinbuttonRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gStatusbarRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gTextRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,GtkEntry,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,GtkTreeView,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gToolbarRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gTreeRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gVarbrowserRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue,gWindowRGtk,guiWidgetsToolkitRGtk2-method}

%%
%% --  svalue<- --
%%
\alias{svalue<--methods}
\alias{svalue<-,guiWidget-method}
\alias{svalue<-,gWidgetRGtk-method} 
%%
%% --  .svalue<- --
%%
\alias{.svalue<--methods}
\alias{.svalue<-,gActionRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gButtonRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gCheckboxRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gCommandlineRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gDroplistRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gEditRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,GtkEntry,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gExpandgroupRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gGraphicsRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gGridRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gGroupRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gImageRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gLabelRGtk,guiWidgetsToolkitRGtk2,ANY,ANY-method}
\alias{.svalue<-,gMenuRGtk,guiWidgetsToolkitRGtk2,ANY,list-method}
\alias{.svalue<-,gMenuRGtk,guiWidgetsToolkitRGtk2,ANY,gMenuRGtk-method}
\alias{.svalue<-,gMenuRGtk,guiWidgetsToolkitRGtk2,ANY,guiWidget-method}
\alias{.svalue<-,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gRadioRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gSliderRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gSpinbuttonRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gStatusbarRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gTextRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gToolbarRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gAddargRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gBivariateRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gModelRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.svalue<-,gWindowRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .leftBracket --
%%
\alias{.leftBracket-methods}
\alias{.leftBracket,gCheckboxRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gCommandlineRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gDfEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gDroplistRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,GtkTreeView,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gLayoutRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gMenuRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gRadioRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,RGtkDataFrame,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gToolbarRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gTreeRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket,gVarbrowserRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .leftBracket<- --
%%
\alias{.leftBracket<--methods}
\alias{.leftBracket<-,gCheckboxRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gDfEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gDroplistRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,GtkTreeView,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gLayoutRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gMenuRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gRadioRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gSliderRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gSpinbuttonRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.leftBracket<-,gToolbarRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  add --
%%
\alias{add-methods}
\alias{add,guiWidget-method}
\alias{add,gWidgetRGtk-method} 
%%
%% --  .add --
%%
\alias{.add-methods}
\alias{.add,guiWidget,guiWidgetsToolkitRGtk2,ANY-method}
\alias{.add,guiWidget,guiWidgetsToolkitRGtk2,guiWidgetORgWidgetRGtkORRGtkObject-method}
\alias{.add,gBasicDialogNoParentRGtk,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.add,gBasicDialogNoParentRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gContainerRGtk,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.add,gContainerRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gDfNotebookRGtk,guiWidgetsToolkitRGtk2,ANY-method}
\alias{.add,gGroupRGtk,guiWidgetsToolkitRGtk2,gGraphicsRGtk-method}
\alias{.add,gGroupRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gGroupRGtk,guiWidgetsToolkitRGtk2,RGtkObject-method}
\alias{.add,gHelpRGtk,guiWidgetsToolkitRGtk2,ANY-method}
\alias{.add,gHelpbrowserRGtk,guiWidgetsToolkitRGtk2,ANY-method}
\alias{.add,gLayoutRGtk,guiWidgetsToolkitRGtk2,ANY-method}
\alias{.add,gLayoutRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gMenuRGtk,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.add,gMenuRGtk,guiWidgetsToolkitRGtk2,gMenuRGtk-method}
\alias{.add,gMenuRGtk,guiWidgetsToolkitRGtk2,list-method}
\alias{.add,gNotebookRGtk,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.add,gNotebookRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gPanedgroupRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gTextRGtk,guiWidgetsToolkitRGtk2,character-method}
\alias{.add,gTextRGtk,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.add,gTextRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gToolbarRGtk,guiWidgetsToolkitRGtk2,list-method}
\alias{.add,gWindowRGtk,guiWidgetsToolkitRGtk2,RGtkObject-method}
\alias{.add,gWindowRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.add,gWindowRGtk,guiWidgetsToolkitRGtk2,gMenuRGtk-method}
\alias{.add,gWindowRGtk,guiWidgetsToolkitRGtk2,gToolbarRGtk-method}
\alias{.add,gWindowRGtk,guiWidgetsToolkitRGtk2,gStatusbarRGtk-method}
\alias{.add,gWidgetRGtk,guiWidgetsToolkitRGtk2,RGtkObject-method}
\alias{.add,gWidgetRGtk,guiWidgetsToolkitRGtk2,try-error-method} 

%%
%% --  addSpace --
%%
\alias{addSpace-methods}
\alias{addSpace,guiWidget-method}
\alias{addSpace,gWidgetRGtk-method} 
%%
%% --  .addSpace --
%%
\alias{.addSpace-methods}
\alias{.addSpace,gContainerRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addSpring --
%%
\alias{addSpring-methods}
\alias{addSpring,guiWidget-method}
\alias{addSpring,gWidgetRGtk-method} 
%%
%% --  .addSpring --
%%
\alias{.addSpring-methods}
\alias{.addSpring,gContainerRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  delete --
%%
\alias{delete-methods}
\alias{delete,guiWidget-method}
\alias{delete,gWidgetRGtk-method} 
%%
%% --  .delete --
%%
\alias{.delete-methods}
\alias{.delete,gContainerRGtk,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.delete,gContainerRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.delete,RGtkObject,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.delete,RGtkObject,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.delete,RGtkObject,guiWidgetsToolkitRGtk2,RGtkObject-method}
\alias{.delete,gWidgetRGtk,guiWidgetsToolkitRGtk2,RGtkObject-method}

\alias{.delete,gWindowRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.delete,gWindowRGtk,guiWidgetsToolkitRGtk2,gMenuRGtk-method}
\alias{.delete,gWindowRGtk,guiWidgetsToolkitRGtk2,gToolbarRGtk-method}
\alias{.delete,gWindowRGtk,guiWidgetsToolkitRGtk2,gStatusbarRGtk-method}

\alias{.delete,gMenuRGtk,guiWidgetsToolkitRGtk2,guiWidget-method}
\alias{.delete,gMenuRGtk,guiWidgetsToolkitRGtk2,gWidgetRGtk-method}
\alias{.delete,gMenuRGtk,guiWidgetsToolkitRGtk2,gMenuRGtk-method}
\alias{.delete,gMenuRGtk,guiWidgetsToolkitRGtk2,list-method}
\alias{.delete,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.delete,gToolbarRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  dispose --
%%
\alias{dispose-methods}
\alias{dispose,guiWidget-method}
\alias{dispose,gWidgetRGtk-method}
\alias{dispose,gTextRGtk-method} 
%%
%% --  .dispose --
%%
\alias{.dispose-methods}
\alias{.dispose,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.dispose,gWindowRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.dispose,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.dispose,gTextRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.dispose,gHelpRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  .insert --
%%
\alias{.insert-methods}
\alias{.insert,gTextRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  visible --
%%
\alias{visible-methods}
\alias{visible,guiWidget-method}
\alias{visible,gWidgetRGtk-method} 
%%
%% --  .visible --
%%
\alias{.visible-methods}
\alias{.visible,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.visible,gBasicDialogNoParentRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.visible,gGridRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  visible<- --
%%
\alias{visible<--methods}
\alias{visible<-,guiWidget-method}
\alias{visible<-,gWidgetRGtk-method} 
\alias{visible<-,gEditRGtk-method} 
%%
%% --  .visible<- --
%%
\alias{.visible<--methods}
\alias{.visible<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.visible<-,gExpandgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.visible<-,gGraphicsRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.visible<-,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.visible<-,gLayoutRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.visible<-,gRadioRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.visible<-,gWindowRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  enabled --
%%
\alias{enabled-methods}
\alias{enabled,guiWidget-method}
\alias{enabled,gWidgetRGtk-method} 
%%
%% --  .enabled --
%%
\alias{.enabled-methods}
\alias{.enabled,GtkWindow,guiWidgetsToolkitRGtk2-method} 
%%
%% --  enabled<- --
%%
\alias{enabled<--methods}
\alias{enabled<-,guiWidget-method}
\alias{enabled<-,gWidgetRGtk-method} 
%%
%% --  .enabled<- --
%%
\alias{.enabled<--methods}
\alias{.enabled<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.enabled<-,gRadioRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.enabled<-,RGtkObject,guiWidgetsToolkitRGtk2-method} 
%%
%% --  size --
%%
\alias{size-methods}
\alias{size,guiWidget-method}
\alias{size,gWidgetRGtk-method} 
%%
%% --  .size --
%%
\alias{.size-methods}
\alias{.size,GtkWindow,guiWidgetsToolkitRGtk2-method}
\alias{.size,gWindowRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.size,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.size,gWidgetRgtk,guiWidgetsToolkitRGtk2-method}
\alias{.size,RGtkObject,guiWidgetsToolkitRGtk2-method} 
%%
%% --  size<- --
%%
\alias{size<--methods}
\alias{size<-,guiWidget-method}
\alias{size<-,gWidgetRGtk-method} 
%%
%% --  .size<- --
%%
\alias{.size<--methods}
\alias{.size<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.size<-,gGroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.size<-,gNotebookRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  focus --
%%
\alias{focus-methods}
\alias{focus,guiWidget-method}
\alias{focus,gWidgetRGtk-method} 
%%
%% --  .focus --
%%
\alias{.focus-methods}
\alias{.focus,gWidgetRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  focus<- --
%%
\alias{focus<--methods}
\alias{focus<-,guiWidget-method}
\alias{focus<-,gWidgetRGtk-method}
\alias{focus<-,RGtkObject-method} 
%%
%% --  .focus<- --
%%
\alias{.focus<--methods}
\alias{.focus<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.focus<-,GtkWindow,guiWidgetsToolkitRGtk2-method}
\alias{.focus<-,RGtkObject,guiWidgetsToolkitRGtk2-method} 
%%
%% --  tooltip<- --
%%
\alias{tooltip<--methods}
\alias{tooltip<-,gWidgetRGtk-method}
\alias{tooltip<-,RGtkObject-method} 
%%
%% --  .tooltip<- --
%%
\alias{.tooltip<--methods}
\alias{.tooltip<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}


%%
%% --  font --
%%
\alias{font-methods}
\alias{font,guiWidget-method}
\alias{font,gWidgetRGtk-method} 
%%
%% --  .font --
%%
\alias{.font-methods}
\alias{.font,GtkWindow,guiWidgetsToolkitRGtk2-method} 
%%
%% --  font<- --
%%
\alias{font<--methods}
\alias{font<-,guiWidget-method}
\alias{font<-,gWidgetRGtk-method} 
\alias{font<-,gButtonRGtk-method} 
%%
%% --  .font<- --
%%
\alias{.font<--methods}
\alias{.font<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.font<-,RGtkObject,guiWidgetsToolkitRGtk2-method}
\alias{.font<-,gButtonRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.font<-,gTextRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  tag --
%%
\alias{tag-methods}
\alias{tag,guiWidget-method}
\alias{tag,gWidgetRGtk-method}
\alias{tag,RGtkObject-method} 
%%
%% --  .tag --
%%
\alias{.tag-methods}
\alias{.tag,guiWidget,guiWidgetsToolkitRGtk2-method}
\alias{.tag,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.tag,RGtkObject,guiWidgetsToolkitRGtk2-method} 
%%
%% --  tag<- --
%%
\alias{tag<--methods}
\alias{tag<-,guiWidget-method}
\alias{tag<-,gWidgetRGtk-method}
\alias{tag<-,RGtkObject-method} 
%%
%% --  .tag<- --
%%
\alias{.tag<--methods}
\alias{.tag<-,guiWidget,guiWidgetsToolkitRGtk2-method}
\alias{.tag<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.tag<-,RGtkObject,guiWidgetsToolkitRGtk2-method} 
%%
%% --  id --
%%
\alias{id-methods}
\alias{id,guiWidget-method}
\alias{id,gWidgetRGtk-method}
\alias{id,RGtkObject-method}
\alias{id,ANY-method}
\alias{id,GtkTreeViewColumn-method} 
%%
%% --  .id --
%%
\alias{.id-methods}
\alias{.id,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.id,RGtkObject,guiWidgetsToolkitRGtk2-method} 
%%
%% --  id<- --
%%
\alias{id<--methods}
\alias{id<-,guiWidget-method}
\alias{id<-,gWidgetRGtk-method}
\alias{id<-,RGtkObject-method}
\alias{id<-,ANY-method}
\alias{id<-,GtkTreeViewColumn-method} 
%%
%% --  .id<- --
%%
\alias{.id<--methods}
\alias{.id<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method} 

%% --  isExtant --
%%
\alias{isExtant-methods}
\alias{isExtant,guiWidget-method}
\alias{isExtant,gWidgetRGtk-method} 
%%
%% --  .isExtant --
%%
\alias{.isExtant-methods}
\alias{.isExtant,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.isExtant,gGridRGtk,guiWidgetsToolkitRGtk2-method} 

%%
%% --  defaultWidget --
%%
\alias{defaultWidget-methods}
\alias{defaultWidget,guiWidget-method}
\alias{defaultWidget,gWidgetRGtk-method}
\alias{defaultWidget,RGtkObject-method} 
\alias{.defaultWidget,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}

%%
%% --  defaultWidget<- --
%%
\alias{defaultWidget<--methods}
\alias{defaultWidget<-,guiWidget-method}
\alias{defaultWidget<-,gWidgetRGtk-method}
\alias{defaultWidget<-,RGtkObject-method} 
\alias{.defaultWidget<-,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.defaultWidget<-,RGtkObject,guiWidgetsToolkitRGtk2-method}

%%
%% --  addhandler --
%%
\alias{addhandler-methods}
\alias{addhandler,guiWidget-method}
\alias{addhandler,gWidgetRGtk-method}
\alias{addhandler,RGtkObject-method}
\alias{addandler,ANY-method}


%%
%% --  blockhandler --
%%
\alias{blockhandler-methods}
\alias{blockhandler,guiWidget-method}
\alias{blockhandler,gWidgetRGtk-method}
\alias{blockhandler,RGtkObject-method}
\alias{blockhandler,ANY-method}
%%
%% --  .blockhandler --
%%
\alias{.blockhandler-methods}
\alias{.blockhandler,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.blockhandler,RGtkObject,guiWidgetsToolkitRGtk2-method}
\alias{.blockhandler,gGridRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.blockhandler,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.blockhandler,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method} 

%%
%% --  unblockhandler --
%%
\alias{unblockhandler-methods}
\alias{unblockhandler,guiWidget-method}
\alias{unblockhandler,gWidgetRGtk-method}
\alias{unblockhandler,RGtkObject-method}
\alias{unblockhandler,ANY-method}
%%
%% --  .unblockhandler --
%%
\alias{.unblockhandler-methods}
\alias{.unblockhandler,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.unblockhandler,RGtkObject,guiWidgetsToolkitRGtk2-method}
\alias{.unblockhandler,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.unblockhandler,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.unblockhandler,gGridRGtk,guiWidgetsToolkitRGtk2-method} 


%%
%% --  removehandler --
%%
\alias{removehandler-methods}
\alias{removehandler,guiWidget-method}
\alias{removehandler,gWidgetRGtk-method}
\alias{removehandler,RGtkObject-method}
\alias{removehandler,GtkTreeViewColumn-method} 
%%
%% --  .removehandler --
%%
\alias{.removehandler-methods}
\alias{.removehandler,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.removehandler,RGtkObject,guiWidgetsToolkitRGtk2-method}
\alias{.removehandler,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.removehandler,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method} 
%%


%%
%% --  addhandler --
%%
\alias{addhandler-methods}
\alias{addhandler,guiWidget-method}
\alias{addhandler,gWidgetRGtk-method}
\alias{addhandler,RGtkObject-method}
\alias{addhandler,GtkTreeViewColumn-method} 
%%
%% --  .addhandler --
%%
\alias{.addhandler-methods}
\alias{.addhandler,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandler,RGtkObject,guiWidgetsToolkitRGtk2-method} 

%% --  addhandlerchanged --
%%
\alias{addhandlerchanged-methods}
\alias{addhandlerchanged,guiWidget-method}
\alias{addhandlerchanged,gWidgetRGtk-method}
\alias{addhandlerchanged,RGtkObject-method}
\alias{addhandlerchanged,ANY-method}
\alias{addhandlerchanged,gSubsetbyRGtk-method}
\alias{addhandlerchanged,GtkTreeViewColumn-method} 
%%
%% --  addHandlerChanged --
%%
\alias{addHandlerChanged-methods}
\alias{addHandlerChanged,guiWidget-method}
\alias{addHandlerChanged,gWidgetRGtk-method}
\alias{addHandlerChanged,RGtkObject-method}
\alias{addHandlerChanged,ANY-method}
\alias{addHandlerChanged,gSubsetbyRGtk-method}
\alias{addHandlerChanged,GtkTreeViewColumn-method} 
%%
%% --  .addhandlerchanged --
%%
\alias{.addhandlerchanged-methods}
\alias{.addhandlerchanged,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gActionRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gButtonRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gCheckboxRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gDroplistRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gExpandgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gGraphicsRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gSubsetbyRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gLabelRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gRadioRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gSliderRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerchanged,gSpinbuttonRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.addhandlerchanged,gTextRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.addhandlerchanged,gTreeRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerkeystroke --
%%
\alias{addhandlerkeystroke-methods}
\alias{addhandlerkeystroke,guiWidget-method}
\alias{addhandlerkeystroke,gWidgetRGtk-method}
\alias{addhandlerkeystroke,RGtkObject-method}
\alias{addhandlerkeystroke,gTextRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .addhandlerkeystroke --
%%
\alias{.addhandlerkeystroke-methods}
\alias{.addhandlerkeystroke,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerkeystroke,gEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerkeystroke,gTextRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerkeystroke,gDroplistRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerclicked --
%%
\alias{addhandlerclicked-methods}
\alias{addhandlerclicked,guiWidget-method}
\alias{addhandlerclicked,gWidgetRGtk-method}
\alias{addhandlerclicked,RGtkObject-method} 
%%
%% --  .addhandlerclicked --
%%
\alias{.addhandlerclicked-methods}
\alias{.addhandlerclicked,RGtkObject,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gButtonRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gCheckboxRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gDroplistRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gGraphicsRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gImageRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gLabelRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerclicked,gRadioRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.addhandlerclicked,gTreeRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerdoubleclick --
%%
\alias{addhandlerdoubleclick-methods}
\alias{addhandlerdoubleclick,guiWidget-method}
\alias{addhandlerdoubleclick,gWidgetRGtk-method}
\alias{addhandlerdoubleclick,RGtkObject-method} 
%%
%% --  .addhandlerdoubleclick --
%%
\alias{.addhandlerdoubleclick-methods}
\alias{.addhandlerdoubleclick,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerdoubleclick,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerdoubleclick,gTreeRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerrightclick --
%%
\alias{addhandlerrightclick-methods}
\alias{addhandlerrightclick,guiWidget-method}
\alias{addhandlerrightclick,gWidgetRGtk-method}
\alias{addhandlerrightclick,RGtkObject-method} 
%%
%% --  .addhandlerrightclick --
%%
\alias{.addhandlerrightclick-methods}
\alias{.addhandlerrightclick,RGtkObject,guiWidgetsToolkitRGtk2-method} 
\alias{.addhandlerrightclick,gWidgetRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerdestroy --
%%
\alias{addhandlerdestroy-methods}
\alias{addhandlerdestroy,guiWidget-method}
\alias{addhandlerdestroy,gWidgetRGtk-method}
\alias{addhandlerdestroy,RGtkObject-method} 
%%
%% --  .addhandlerdestroy --
%%
\alias{.addhandlerdestroy-methods}
\alias{.addhandlerdestroy,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerdestroy,gWindowRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerexpose --
%%
\alias{addhandlerexpose-methods}
\alias{addhandlerexpose,guiWidget-method}
\alias{addhandlerexpose,gWidgetRGtk-method}
\alias{addhandlerexpose,RGtkObject-method} 
%%
%% --  .addhandlerexpose --
%%
\alias{.addhandlerexpose-methods}
\alias{.addhandlerexpose,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerexpose,gComponentRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerexpose,gGraphicsRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerexpose,gNotebookRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerunrealize --
%%
\alias{addhandlerunrealize-methods}
\alias{addhandlerunrealize,guiWidget-method}
\alias{addhandlerunrealize,gWidgetRGtk-method}
\alias{addhandlerunrealize,RGtkObject-method} 
%%
%% --  .addhandlerunrealize --
%%
\alias{.addhandlerunrealize-methods}
\alias{.addhandlerunrealize,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlerunrealize,gWindowRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addhandlerblur --
%%
\alias{addhandlerblur-methods}
\alias{addhandlerblur,guiWidget-method}
\alias{addhandlerblur,gWidgetRGtk-method}
\alias{addhandlerblur,RGtkObject-method} 
%%
%% --  .addhandlerblur --
%%
\alias{.addhandlerblur-methods}
\alias{.addhandlerblur,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  addhandlerfocus --
%%
\alias{addhandlerfocus-methods}
\alias{addhandlerfocus,guiWidget-method}
\alias{addhandlerfocus,gWidgetRGtk-method}
\alias{addhandlerfocus,RGtkObject-method} 
%%
%% --  .addhandlerfocus --
%%
\alias{.addhandlerfocus-methods}
\alias{.addhandlerfocus,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  addhandleridle --
%%
\alias{addhandleridle-methods}
\alias{addhandleridle,guiWidget-method}
\alias{addhandleridle,gWidgetRGtk-method}
\alias{addhandleridle,RGtkObject-method} 
%%
%% --  .addhandleridle --
%%
\alias{.addhandleridle-methods}
\alias{.addhandleridle,gWidgetRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  addpopupmenu --
%%
\alias{addpopupmenu-methods}
\alias{addpopupmenu,guiWidget-method}
\alias{addpopupmenu,gWidgetRGtk-method}
\alias{addpopupmenu,RGtkObject-method}
%%
%% --  .addpopupmenu --
%%
\alias{.addpopupmenu-methods}
\alias{.addpopupmenu,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addpopupmenu,gButtonRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.addpopupmenu,gLabelRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  add3rdmousepopupmenu --
%%
\alias{add3rdmousepopupmenu-methods}
\alias{add3rdmousepopupmenu,guiWidget-method}
\alias{add3rdmousepopupmenu,gWidgetRGtk-method}
\alias{add3rdmousepopupmenu,RGtkObject-method} 
%%
%% --  .add3rdmousepopupmenu --
%%
\alias{.add3rdmousepopupmenu-methods}
\alias{.add3rdmousepopupmenu,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.add3rdmousepopupmenu,RGtkObject,guiWidgetsToolkitRGtk2-method} 
\alias{.add3rdmousepopupmenu,gLabelRGtk,guiWidgetsToolkitRGtk2-method}
 
%%
%% --  addmousemotion --
%%
\alias{addhandlermousemotion-methods}
\alias{addhandlermousemotion,guiWidget-method}
\alias{addhandlermousemotion,gWidgetRGtk-method}
\alias{addhandlermousemotion,RGtkObject-method} 
%%
%% --  .add3rdmousepopupmenu --
%%
\alias{.addhandlermousemotion-methods}
\alias{.addhandlermousemotion,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.addhandlermousemotion,RGtkObject,guiWidgetsToolkitRGtk2-method} 
\alias{.addhandlermousemotion,gLabelRGtk,guiWidgetsToolkitRGtk2-method}

%%
%% --  adddropsource --
%%
\alias{adddropsource-methods}
\alias{adddropsource,guiWidget-method}
\alias{adddropsource,gWidgetRGtk-method}
\alias{adddropsource,RGtkObject-method} 
%%
%% --  .adddropsource --
%%
\alias{.adddropsource-methods}
\alias{.adddropsource,RGtkObject,guiWidgetsToolkitRGtk2-method}
\alias{.adddropsource,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.adddropsource,gLabelRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.adddropsource,gEditRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  adddropmotion --
%%
\alias{adddropmotion-methods}
\alias{adddropmotion,guiWidget-method}
\alias{adddropmotion,gWidgetRGtk-method}
\alias{adddropmotion,RGtkObject-method} 
%%
%% --  .adddropmotion --
%%
\alias{.adddropmotion-methods}
\alias{.adddropmotion,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.adddropmotion,RGtkObject,guiWidgetsToolkitRGtk2-method} 
%%
%% --  adddroptarget --
%%
\alias{adddroptarget-methods}
\alias{adddroptarget,guiWidget-method}
\alias{adddroptarget,gWidgetRGtk-method}
\alias{adddroptarget,RGtkObject-method} 
\alias{adddroptarget,gEditRGtk-method} 
%%
%% --  .adddroptarget --
%%
\alias{.adddroptarget-methods}
\alias{.adddroptarget,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.adddroptarget,RGtkObject,guiWidgetsToolkitRGtk2-method}
\alias{.adddroptarget,gEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.adddroptarget,gImageRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.adddroptarget,gLabelRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  .galert --
%%
\alias{.galert-methods}
\alias{.galert,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gmessage --
%%
\alias{.gmessage-methods}
\alias{.gmessage,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .ginput --
%%
\alias{.ginput-methods}
\alias{.ginput,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gconfirm --
%%
\alias{.gconfirm-methods}
\alias{.gconfirm,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gbasicdialog --
%%
\alias{.gbasicdialog-methods}
\alias{.gbasicdialog,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .gbasicdialognoparent --
%%
\alias{.gbasicdialognoparent-methods}
\alias{.gbasicdialognoparent,guiWidgetsToolkitRGtk2-method} 
%%
%% --  update --
%%
\alias{update-methods}
\alias{update,ANY-method}
\alias{update,guiWidget-method}
\alias{update,gWidgetRGtk-method}
\alias{update,gSubsetbyRGtk-method}
\alias{update,gTreeRGtk-method} 
%%
%% --  .update --
%%
\alias{.update-methods}
\alias{.update,gComponentRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.update,gVarbrowserRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.update,gTreeRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .length --
%%
\alias{.length-methods}
\alias{.length,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gCheckboxgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gCheckboxgroupTableRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gDfEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gDroplistRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gHelpRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.length,gRadioRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  dim --
%%
\alias{dim-methods}
\alias{dim,ANY-method}
\alias{dim,guiWidget-method}
\alias{dim,gWidgetRGtk-method} 
%%
%% --  .dim --
%%
\alias{.dim-methods}
\alias{.dim,gDfEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.dim,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.dim,gLayoutRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  .dimnames --
%%
\alias{.dimnames-methods}
\alias{.dimnames,gGridRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.dimnames,gDfEditRGtk,guiWidgetsToolkitRGtk2-method} 

%%
%% --  .dimnames<- --
%%
\alias{.dimnames<--methods}
\alias{.dimnames<-,gGridRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.dimnames<-,gDfEditRGtk,guiWidgetsToolkitRGtk2-method} 
%%
%% --  names --
%%
\alias{names-methods}
\alias{names,ANY-method}
\alias{names,guiWidget-method}
\alias{names,gWidgetRGtk-method} 
\alias{names,gFrameRGtk-method}
\alias{names,gExpandgroupRGtk-method} 
%%
%% --  .names --
%%
\alias{.names-methods}
\alias{.names,gDfEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.names,gExpandgroupRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.names,gFrameRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.names,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.names,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% --  names<- --
%%
\alias{names<--methods}
\alias{names<-,ANY-method}
\alias{names<-,guiWidget-method}
\alias{names<-,gWidgetRGtk-method} 
%%
%% --  .names<- --
%%
\alias{.names<--methods}
\alias{.names<-,gDfEditRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.names<-,gExpandgroupRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.names<-,gFrameRGtk,guiWidgetsToolkitRGtk2-method} 
\alias{.names<-,gGridRGtk,guiWidgetsToolkitRGtk2-method}
\alias{.names<-,gNotebookRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% -- .getToolkitWidget --
\alias{.getToolkitWidget,gWidgetRGtk,guiWidgetsToolkitRGtk2-method}
%%
%% -- [<-
\alias{[<-,gWidgetRGtk-method}
\alias{[<-,gCheckboxRGtk-method}
\alias{[<-,gCheckboxgroupRGtk-method}
\alias{[<-,gCheckboxgroupTableRGtk-method}
\alias{[<-,gDfEditRGtk-method}
\alias{[<-,gDroplistRGtk-method}
\alias{[<-,gEditRGtk-method}
\alias{[<-,GtkTreeView-method}
\alias{[<-,gGridRGtk-method}
\alias{[<-,gLayoutRGtk-method}
\alias{[<-,gMenuRGtk-method}
\alias{[<-,gNotebookRGtk-method}
\alias{[<-,gRadioRGtk-method}
\alias{[<-,gSliderRGtk-method}
\alias{[<-,gSpinbuttonRGtk-method}
\alias{[<-,gToolbarRGtk-method}
%% -- dimnames
\alias{dimnames,gWidgetRGtk-method}
\alias{dimnames<-,gWidgetRGtk-method}
\alias{length,gWidgetRGtk-method}
\alias{length,gDroplistRGtk-method}
\alias{length,gSubsetbyRGtk-method}

%%% Classes from  RGTk2 upgraded to S4 classes
\alias{AtkNoOpObjectFactory-class}
\alias{AtkObjectFactory-class}
\alias{AtkRelationSet-class}
\alias{AtkStateSet-class}
\alias{GBoxed-class}
\alias{GObject-class}
\alias{GScanner-class}
\alias{GdkDragContext-class}
\alias{GdkPixbufLoader-class}
\alias{GdkRegion-class}
\alias{GtkAboutDialog-class}
\alias{GtkAccelGroup-class}
\alias{GtkAccelLabel-class}
\alias{GtkAction-class}
\alias{GtkActionGroup-class}
\alias{GtkAdjustment-class}
\alias{GtkAlignment-class}
\alias{GtkArrow-class}
\alias{GtkAspectFrame-class}
\alias{GtkBin-class}
\alias{GtkBox-class}
\alias{GtkButton-class}
\alias{GtkButtonBox-class}
\alias{GtkCList-class}
\alias{GtkCTree-class}
\alias{GtkCalendar-class}
\alias{GtkCellRenderer-class}
\alias{GtkCellRendererCombo-class}
\alias{GtkCellRendererPixbuf-class}
\alias{GtkCellRendererProgress-class}
\alias{GtkCellRendererText-class}
\alias{GtkCellRendererToggle-class}
\alias{GtkCellView-class}
\alias{GtkCheckButton-class}
\alias{GtkCheckMenuItem-class}
\alias{GtkColorButton-class}
\alias{GtkColorSelection-class}
\alias{GtkColorSelectionDialog-class}
\alias{GtkCombo-class}
\alias{GtkComboBox-class}
\alias{GtkComboBoxEntry-class}
\alias{GtkContainer-class}
\alias{GtkCurve-class}
\alias{GtkDialog-class}
\alias{GtkDrawingArea-class}
\alias{GtkEntry-class}
\alias{GtkEntryCompletion-class}
\alias{GtkEventBox-class}
\alias{GtkExpander-class}
\alias{GtkFileFilter-class}
\alias{GtkFileSelection-class}
\alias{GtkFixed-class}
\alias{GtkFontButton-class}
\alias{GtkFontSelection-class}
\alias{GtkFontSelectionDialog-class}
\alias{GtkFrame-class}
\alias{GtkGammaCurve-class}
\alias{GtkHBox-class}
\alias{GtkHButtonBox-class}
\alias{GtkHPaned-class}
\alias{GtkHRuler-class}
\alias{GtkHScale-class}
\alias{GtkHScrollbar-class}
\alias{GtkHSeparator-class}
\alias{GtkHandleBox-class}
\alias{GtkIMContext-class}
\alias{GtkIMContextSimple-class}
\alias{GtkIMMulticontext-class}
\alias{GtkIconFactory-class}
\alias{GtkIconSet-class}
\alias{GtkIconSource-class}
\alias{GtkIconTheme-class}
\alias{GtkIconView-class}
\alias{GtkImage-class}
\alias{GtkImageMenuItem-class}
\alias{GtkInputDialog-class}
\alias{GtkInvisible-class}
\alias{GtkItem-class}
\alias{GtkLabel-class}
\alias{GtkLayout-class}
\alias{GtkList-class}
\alias{GtkListItem-class}
\alias{GtkMenu-class}
\alias{GtkMenuBar-class}
\alias{GtkMenuItem-class}
\alias{GtkMenuShell-class}
\alias{GtkMisc-class}
\alias{GtkNotebook-class}
\alias{GtkObject-class}
\alias{GtkOptionMenu-class}
\alias{GtkPaned-class}
\alias{GtkProgress-class}
\alias{GtkProgressBar-class}
\alias{GtkRadioAction-class}
\alias{GtkRadioButton-class}
\alias{GtkRange-class}
\alias{GtkRcStyle-class}
\alias{GtkRuler-class}
\alias{GtkScale-class}
\alias{GtkScrollbar-class}
\alias{GtkScrolledWindow-class}
\alias{GtkSeparator-class}
\alias{GtkSeparatorMenuItem-class}
\alias{GtkSeparatorToolItem-class}
\alias{GtkSizeGroup-class}
\alias{GtkSocket-class}
\alias{GtkSpinButton-class}
\alias{GtkStatusbar-class}
\alias{GtkStyle-class}
\alias{GtkTable-class}
\alias{GtkTearoffMenuItem-class}
\alias{GtkTextAttributes-class}
\alias{GtkTextBuffer-class}
\alias{GtkTextChildAnchor-class}
\alias{GtkTextTag-class}
\alias{GtkTextTagTable-class}
\alias{GtkTextView-class}
\alias{GtkTipsQuery-class}
\alias{GtkToggleAction-class}
\alias{GtkToggleButton-class}
\alias{GtkToggleToolButton-class}
\alias{GtkToolButton-class}
\alias{GtkToolItem-class}
\alias{GtkToolbar-class}
\alias{GtkTooltips-class}
\alias{GtkTreeModelSort-class}
\alias{GtkTreePath-class}
\alias{GtkTreeView-class}
\alias{GtkTreeViewColumn-class}
\alias{GtkUIManager-class}
\alias{GtkVBox-class}
\alias{GtkVButtonBox-class}
\alias{GtkVPaned-class}
\alias{GtkVRuler-class}
\alias{GtkVScale-class}
\alias{GtkVScrollbar-class}
\alias{GtkVSeparator-class}
\alias{GtkViewport-class}
\alias{GtkWidget-class}
\alias{GtkWindow-class}
\alias{GtkWindowGroup-class}
\alias{PangoAttrList-class}
\alias{PangoCairoFcFontMap-class}
\alias{PangoCoverage-class}
\alias{PangoFcFontMap-class}
\alias{PangoFontDescription-class}
\alias{PangoFontMap-class}
\alias{PangoGlyphString-class}
\alias{PangoItem-class}
\alias{RGtkDataFrame-class}



\title{Undocumented functions in gWidgetsRGtk}
\description{
	The gWidgetsRGtk package implements the gWidgets API. In
	gWidgets the primary constructors and methods are
	documented. The basic idea is that \code{function} in gWidgets
	calls \code{.function} in gWidgetsRGtk. Hence the many
	functions here.
}
\keyword{interface}% at least one, from doc/KEYWORDS
\keyword{internal}