\name{gWidgetsRGtk2-misc}
\alias{gWidgetsRGtk2-misc}
\alias{Paste}
\alias{stripWhiteSpace}
\alias{rpel}
\alias{str1}
\alias{str2}
\alias{untaintName}
\alias{stockIconFromClass}
\alias{.stockIconFromClass-methods}
\alias{.stockIconFromClass,guiWidgetsToolkitRGtk2-method} 
\alias{stockIconFromObject}
\alias{.stockIconFromObject-methods}
\alias{.stockIconFromObject,guiWidgetsToolkitRGtk2-method} 
\alias{Timestamp}
\alias{Timestamp<-}
       
\title{Miscellaneous functions in gWidgetsRGtk} \description{ These
functions are hardly worth documenting. They are used by pmg, but are
not part of the gWidgets API, nor meant for general consumption.
}
\keyword{interface}% at least one, from doc/KEYWORDS
